+++ src/checks.rs
@@ -15,6 +15,7 @@ mod unordered_key;
 // This trait is used for checks which needs to know of only a single line
 trait Check {
     fn run(&mut self, line: &LineEntry) -> Option<Warning>;
+    fn name(&self) -> &str;
 }
 
 // Checklist for checks which needs to know of only a single line
@@ -33,8 +34,24 @@ fn checklist() -> Vec<Box<dyn Check>> {
     ]
 }
 
-pub fn run(lines: Vec<LineEntry>) -> Vec<Warning> {
+pub fn available_check_names() -> Vec<String> {
+    let mut names: Vec<String> = checklist()
+        .iter()
+        .map(|check| check.name().to_string())
+        .collect();
+    names.push(
+        ending_blank_line::EndingBlankLineChecker::default()
+            .name()
+            .to_string(),
+    );
+    names.sort();
+    names
+}
+
+pub fn run(lines: Vec<LineEntry>, skip_checks: &[&str]) -> Vec<Warning> {
     let mut checks = checklist();
+    checks.retain(|c| !skip_checks.contains(&c.name()));
+
     let mut warnings: Vec<Warning> = Vec::new();
 
     for line in &lines {
@@ -51,8 +68,11 @@ pub fn run(lines: Vec<LineEntry>) -> Vec<Warning> {
 
     if let Some(last_line) = lines.last() {
         let mut ending_line_checker = ending_blank_line::EndingBlankLineChecker::default();
-        if let Some(warning) = ending_line_checker.run(last_line) {
-            warnings.push(warning);
+
+        if !skip_checks.contains(&ending_line_checker.name()) {
+            if let Some(warning) = ending_line_checker.run(last_line) {
+                warnings.push(warning);
+            }
         }
     }
 
@@ -64,10 +84,13 @@ mod tests {
     use super::*;
     use std::path::PathBuf;
 
-    fn blank_entry(number: usize) -> LineEntry {
+    fn blank_line_entry(number: usize) -> LineEntry {
         LineEntry {
             number,
-            file_path: PathBuf::from(".env"),
+            file: FileEntry {
+                path: PathBuf::from(".env"),
+                file_name: ".env".to_string(),
+            },
             raw_string: String::from("\n"),
         }
     }
@@ -75,7 +98,10 @@ mod tests {
     fn line_entry(number: usize, str: &str) -> LineEntry {
         LineEntry {
             number,
-            file_path: PathBuf::from(".env"),
+            file: FileEntry {
+                path: PathBuf::from(".env"),
+                file_name: ".env".to_string(),
+            },
             raw_string: String::from(str),
         }
     }
@@ -84,32 +110,36 @@ mod tests {
     fn run_with_empty_vec_test() {
         let empty: Vec<LineEntry> = Vec::new();
         let expected: Vec<Warning> = Vec::new();
+        let skip_checks: Vec<&str> = Vec::new();
 
-        assert_eq!(expected, run(empty));
+        assert_eq!(expected, run(empty, &skip_checks));
     }
 
     #[test]
     fn run_with_empty_line_test() {
-        let lines: Vec<LineEntry> = vec![blank_entry(1)];
+        let lines: Vec<LineEntry> = vec![blank_line_entry(1)];
         let expected: Vec<Warning> = Vec::new();
+        let skip_checks: Vec<&str> = Vec::new();
 
-        assert_eq!(expected, run(lines));
+        assert_eq!(expected, run(lines, &skip_checks));
     }
 
     #[test]
     fn run_with_comment_line_test() {
-        let lines: Vec<LineEntry> = vec![line_entry(1, "# Comment = 'Value'"), blank_entry(2)];
+        let lines: Vec<LineEntry> = vec![line_entry(1, "# Comment = 'Value'"), blank_line_entry(2)];
         let expected: Vec<Warning> = Vec::new();
+        let skip_checks: Vec<&str> = Vec::new();
 
-        assert_eq!(expected, run(lines));
+        assert_eq!(expected, run(lines, &skip_checks));
     }
 
     #[test]
     fn run_with_valid_line_test() {
-        let lines: Vec<LineEntry> = vec![line_entry(1, "FOO=BAR"), blank_entry(2)];
+        let lines: Vec<LineEntry> = vec![line_entry(1, "FOO=BAR"), blank_line_entry(2)];
         let expected: Vec<Warning> = Vec::new();
+        let skip_checks: Vec<&str> = Vec::new();
 
-        assert_eq!(expected, run(lines));
+        assert_eq!(expected, run(lines, &skip_checks));
     }
 
     #[test]
@@ -121,10 +151,11 @@ mod tests {
                 "KeyWithoutValue: The FOO key should be with a value or have an equal sign",
             ),
         );
-        let lines: Vec<LineEntry> = vec![line, blank_entry(2)];
+        let lines: Vec<LineEntry> = vec![line, blank_line_entry(2)];
         let expected: Vec<Warning> = vec![warning];
+        let skip_checks: Vec<&str> = Vec::new();
 
-        assert_eq!(expected, run(lines));
+        assert_eq!(expected, run(lines, &skip_checks));
     }
 
     #[test]
@@ -136,7 +167,44 @@ mod tests {
         );
         let lines: Vec<LineEntry> = vec![line];
         let expected: Vec<Warning> = vec![warning];
+        let skip_checks: Vec<&str> = Vec::new();
 
-        assert_eq!(expected, run(lines));
+        assert_eq!(expected, run(lines, &skip_checks));
+    }
+
+    #[test]
+    fn skip_one_check() {
+        let line1 = line_entry(1, "FOO");
+        let line2 = line_entry(2, "1FOO");
+        let warning = Warning::new(
+            line2.clone(),
+            String::from("LeadingCharacter: Invalid leading character detected"),
+        );
+        let lines: Vec<LineEntry> = vec![line1, line2, blank_line_entry(3)];
+        let expected: Vec<Warning> = vec![warning];
+        let skip_checks: Vec<&str> = vec!["KeyWithoutValue"];
+
+        assert_eq!(expected, run(lines, &skip_checks));
+    }
+
+    #[test]
+    fn skip_all_checks() {
+        let line = line_entry(1, "FOO");
+        let lines: Vec<LineEntry> = vec![line];
+        let expected: Vec<Warning> = Vec::new();
+        let skip_checks: Vec<&str> = vec!["KeyWithoutValue", "EndingBlankLine"];
+
+        assert_eq!(expected, run(lines, &skip_checks));
+    }
+
+    #[test]
+    fn check_name_list() {
+        let available_check_names = available_check_names();
+        for check in checklist() {
+            let check_name = check.name();
+            assert!(available_check_names.contains(&check_name.to_string()));
+        }
+        let ending_blank_line = ending_blank_line::EndingBlankLineChecker::default();
+        assert!(available_check_names.contains(&ending_blank_line.name().to_string()));
     }
 }