+++ src/checks/extra_blank_line.rs
@@ -1,19 +1,19 @@
 use crate::checks::Check;
 use crate::common::*;
 
-pub(crate) struct ExtraBlankLineChecker {
-    template: &'static str,
-    name: &'static str,
+pub(crate) struct ExtraBlankLineChecker<'a> {
+    template: &'a str,
+    name: &'a str,
     last_blank_number: Option<usize>,
 }
 
-impl ExtraBlankLineChecker {
+impl ExtraBlankLineChecker<'_> {
     fn message(&self) -> String {
         return format!("{}: {}", self.name, self.template);
     }
 }
 
-impl Default for ExtraBlankLineChecker {
+impl Default for ExtraBlankLineChecker<'_> {
     fn default() -> Self {
         Self {
             name: "ExtraBlankLine",
@@ -23,20 +23,26 @@ impl Default for ExtraBlankLineChecker {
     }
 }
 
-impl Check for ExtraBlankLineChecker {
+impl Check for ExtraBlankLineChecker<'_> {
     fn run(&mut self, line: &LineEntry) -> Option<Warning> {
         if !line.is_empty() {
             return None;
         }
 
-        if let Some(last_blank_number) = self.last_blank_number {
-            if last_blank_number + 1 == line.number {
-                return Some(Warning::new(line.clone(), self.message()));
-            }
-        }
+        let is_extra = self
+            .last_blank_number
+            .map_or(false, |n| n + 1 == line.number);
         self.last_blank_number = Some(line.number);
 
+        if is_extra {
+            return Some(Warning::new(line.clone(), self.message()));
+        }
+
         None
+    }
+
+    fn name(&self) -> &str {
+        self.name
     }
 }
 
@@ -52,7 +58,10 @@ mod tests {
             let (content, message) = *assert;
             let line = LineEntry {
                 number: i + 1,
-                file_path: PathBuf::from(".env"),
+                file: FileEntry {
+                    path: PathBuf::from(".env"),
+                    file_name: ".env".to_string(),
+                },
                 raw_string: String::from(content),
             };
             let expected = message.map(|msg| Warning::new(line.clone(), String::from(msg)));
@@ -80,6 +89,19 @@ mod tests {
         let asserts = vec![
             ("A=B", None),
             ("", None),
+            ("", Some("ExtraBlankLine: Extra blank line detected")),
+            ("C=D", None),
+        ];
+
+        run_asserts(asserts);
+    }
+
+    #[test]
+    fn three_blank_lines() {
+        let asserts = vec![
+            ("A=B", None),
+            ("", None),
+            ("", Some("ExtraBlankLine: Extra blank line detected")),
             ("", Some("ExtraBlankLine: Extra blank line detected")),
             ("C=D", None),
         ];